+++ tests/testthat/test-plot_energy.R
@@ -1,0 +1,19 @@
+testthat::context("test-plot_energy")
+
+pattern_reconstruction <- shar::reconstruct_pattern(pattern = shar::species_a,
+                                                    n_random = 3,
+                                                    max_runs = 100,
+                                                    verbose = FALSE)
+
+################################################################################
+
+testthat::test_that("plot_energy returns plot", {
+
+  testthat::expect_null(shar::plot_energy(pattern_reconstruction))
+})
+
+testthat::test_that("plot_energy uses colours", {
+
+  testthat::expect_null(shar::plot_energy(pattern_reconstruction,
+                                          col = c("blue", "green", "red")))
+})